requirements.txt
streamlit>=1.37
tensorflow>=2.17,<2.18  
torch>=2.3.0
torchvision>=0.18.0
timm>=1.0.0              
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.26.0
